Reformat Source Code
Reformat files and project structure to allow for better separation of concerns and provide room for future expansion
Grouping and creating objects with similar functionality
