--- lambda-backend-challenge/src/lambdas/breeds-get.test.ts
@@ -1,6 +1,0 @@
-import fetch from 'node-fetch'
-import { handler } from './breeds-get'
-
-const mockedFetch: jest.Mock = fetch as any
-
-jest.mock('node-fetch')
