+++ lambda-backend-challenge/src/lambdas/index.ts
@@ -1,2 +1,2 @@
-export { handler as randomGet } from './random-get'
-export { handler as breedsGet } from './breeds-get'
+export { handler as randomGet } from './modules/random-get'
+export { handler as breedsGet } from './modules/breeds-get'
