+++ lambda-backend-challenge/src/lambdas/interfaces/breeds-list.ts
@@ -1,0 +1,5 @@
+
+export interface BreedsList {
+  message: string
+  status: string
+}
