+++ lambda-backend-challenge/src/lambdas/interfaces/random-dog.ts
@@ -1,0 +1,5 @@
+
+export interface RandomDog {
+  message: string
+  status: string
+}
