+++ lambda-backend-challenge/src/lambdas/tests/random-get.test.ts
@@ -1,5 +1,5 @@
 import fetch from 'node-fetch'
-import { handler } from './random-get'
+import { handler } from '../modules/random-get'
 
 const mockedFetch: jest.Mock = fetch as any
